DB context fixes. DB migrations. Startup updated to usage dev/prod DB.

+++ src/Tti.Estate.Data/Entities/Transaction.cs
@@ -28,9 +28,9 @@ namespace Tti.Estate.Data.Entities
 
         public decimal Amount { get; set; }
 
-        public decimal CompanyPercent { get; set; }
+        public byte CompanyPercent { get; set; }
 
-        public decimal UserPercent { get; set; }
+        public byte UserPercent { get; set; }
 
         public string Description { get; set; }
